+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased](https://github.com/Stadly/PasswordPolice/compare/v0.19.0...HEAD)
 
 ### Added
-- Nothing
+- Character tree cutter separating a string root of specified length from the rest of the tree.
 
 ### Changed
 - Nothing